added: play dvd button in my videos
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@552 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ guilib/GUIWindowManager.cpp
@@ -68,7 +68,7 @@ void CGUIWindowManager::PreviousWindow()
 	if (m_iActiveWindow >=0)
 	{
     
-    if (m_iActiveWindow >=0 && m_iActiveWindow < m_vecWindows.size())
+    if (m_iActiveWindow >=0 && m_iActiveWindow < (int)m_vecWindows.size())
     {
 		  CGUIWindow* pWindow=m_vecWindows[m_iActiveWindow];
 		  iPrevActiveWindowID = pWindow->GetPreviousWindowID();
